Display immediate icon color change on like/dislike of photo
Previously, a manual refresh of the browswer was required to display
the color change of the heart icon (after user liked/disliked a photo).
This was due to a mismatch of snakecase vs camelcase between backend
and frontend, respectively. This problem was resolved by adding test
coverage and ensuring that the likes jbuilder pattern matches the
photo jbuilder pattern.

This PR also includes alignment of redux code to modern conventions.